src/main/webapp/app/entities/utilisation/delete/utilisation-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { UtilisationService } from '../service/utilisation.service';

import { UtilisationDeleteDialogComponent } from './utilisation-delete-dialog.component';

describe('Utilisation Management Delete Component', () => {
  let comp: UtilisationDeleteDialogComponent;
  let fixture: ComponentFixture<UtilisationDeleteDialogComponent>;
  let service: UtilisationService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule],
      declarations: [UtilisationDeleteDialogComponent],
      providers: [NgbActiveModal],
    })
      .overrideTemplate(UtilisationDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(UtilisationDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(UtilisationService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      })
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
